Bug 833624. Bump up size of int's in mOrder to allow more than 256 CSS properties. r=dbaron

+++ layout/style/Declaration.h
@@ -230,7 +230,7 @@ public:
   size_t SizeOfIncludingThis(nsMallocSizeOfFun aMallocSizeOf) const;
 
 private:
-  nsAutoTArray<uint8_t, 8> mOrder;
+  nsAutoTArray<uint16_t, 8> mOrder;
 
   // never null, except while expanded, or before the first call to
   // InitializeEmpty or CompressFrom.
